+++ upgrade/0.0.7/upgrade-o.sh
@@ -13,6 +13,9 @@ function upgrade_mmgen {
 }
 
 VER='btc_bch'
+echo "Skipping MMGen upgrade to version '$VER'"
+echo "The 'mmlive-upgrade' utility will install the latest version automatically"
+exit 0
 [ -d ~mmgen/src/mmgen-$VER ] || upgrade_mmgen
 
 exit 0